Hash changes
Hashes used to be calculated on the album art thumbnail, and the full-size album art. This used two different hashing methods. Now, hashing is only done on the full image, and uses the average hash method. The other hashing methods were removed. This can be further improved by getting the image hash from the genius album picture, and using that to search for a higher quality version