Massive accuracy improvements🎉💯
Google Images album cover downloader used to fail every time. Now it works, by using a headless browser and parsing with beautifulsoup.

Amazon CD album downloader source has been removed. It was not that accurate, and the other sources should be fine on their own.

All CoverSourceResult children have been removed, due to a simplifications in CoverSourceResult. The image is downloaded, and metadata found, in the constructor